King Escape
Alice and Bob are playing chess on a huge chessboard with dimensions n×n. Alice has a single piece left — a queen, located at (ax,ay), while Bob has only the king standing at (bx,by). Alice thinks that as her queen is dominating the chessboard, victory is hers.

But Bob has made a devious plan to seize the victory for himself — he needs to march his king to (cx,cy) in order to claim the victory for himself. As Alice is distracted by her sense of superiority, she no longer moves any pieces around, and it is only Bob who makes any turns.

Bob will win if he can move his king from (bx,by) to (cx,cy) without ever getting in check. Remember that a king can move to any of the 8 adjacent squares. A king is in check if it is on the same rank (i.e. row), file (i.e. column), or diagonal as the enemy queen.

Find whether Bob can win or not.

https://codeforces.com/problemset/problem/1033/A