Make it Increasing
You are given an array a consisting of n positive integers, and an array b, with length n. Initially bi=0 for each 1≤i≤n.

In one move you can choose an integer i (1≤i≤n), and add ai to bi or subtract ai from bi. What is the minimum number of moves needed to make b increasing (that is, every element is strictly greater than every element before it)?

https://codeforces.com/problemset/problem/1667/A