Twist the Permutation
Petya got an array a of numbers from 1 to n, where a[i]=i.

He performed n operations sequentially. In the end, he received a new state of the a array.

At the i-th operation, Petya chose the first i elements of the array and cyclically shifted them to the right an arbitrary number of times (elements with indexes i+1 and more remain in their places). One cyclic shift to the right is such a transformation that the array a=[a1,a2,…,an] becomes equal to the array a=[ai,a1,a2,…,ai−2,ai−1,ai+1,ai+2,…,an].

For example, if a=[5,4,2,1,3] and i=3 (that is, this is the third operation), then as a result of this operation, he could get any of these three arrays:

a=[5,4,2,1,3] (makes 0 cyclic shifts, or any number that is divisible by 3);
a=[2,5,4,1,3] (makes 1 cyclic shift, or any number that has a remainder of 1 when divided by 3);
a=[4,2,5,1,3] (makes 2 cyclic shifts, or any number that has a remainder of 2 when divided by 3).
Let's look at an example. Let n=6, i.e. initially a=[1,2,3,4,5,6]. A possible scenario is described below.

i=1: no matter how many cyclic shifts Petya makes, the array a does not change.
i=2: let's say Petya decided to make a 1 cyclic shift, then the array will look like a=[2,1,3,4,5,6].
i=3: let's say Petya decided to make 1 cyclic shift, then the array will look like a=[3,2,1,4,5,6].
i=4: let's say Petya decided to make 2 cyclic shifts, the original array will look like a=[1,4,3,2,5,6].
i=5: let's say Petya decided to make 0 cyclic shifts, then the array won't change.
i=6: let's say Petya decided to make 4 cyclic shifts, the array will look like a=[3,2,5,6,1,4].
You are given a final array state a after all n operations. Determine if there is a way to perform the operation that produces this result. In this case, if an answer exists, print the numbers of cyclical shifts that occurred during each of the n operations.

https://codeforces.com/problemset/problem/1650/D