Party
A club plans to hold a party and will invite some of its n members. The n members are identified by the numbers 1,2,…,n. If member i is not invited, the party will gain an unhappiness value of ai.

There are m pairs of friends among the n members. As per tradition, if both people from a friend pair are invited, they will share a cake at the party. The total number of cakes eaten will be equal to the number of pairs of friends such that both members have been invited.

However, the club's oven can only cook two cakes at a time. So, the club demands that the total number of cakes eaten is an even number.

What is the minimum possible total unhappiness value of the party, respecting the constraint that the total number of cakes eaten is even?

https://codeforces.com/problemset/problem/1711/B